requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
openai==0.27.4
Pygments==2.14.0
python-dotenv==1.0.0
requests==2.28.2
rich==12.6.0
shellingham==1.5.0.post1
tqdm==4.65.0
typer==0.7.0
urllib3==1.26.15
wincertstore==0.2
yarl==1.8.2
